Fixed point

+++ src/filter_if.cpp
@@ -103,7 +103,7 @@ fm_call_def *fm_comp_filter_if_stream_call(fm_comp_def_cl comp_cl,
 
 void fm_comp_filter_if_queuer(size_t idx, fm_call_ctx_t *ctx) {
   auto *comp_cl = (filter_if_comp_cl *)ctx->comp;
-  if (comp_cl->idx == idx) {
+  if (comp_cl->idx == (int)idx) {
     comp_cl->updated = true;
   }
 }
